requirements.txt
flask
gunicorn
azure-storage-blob>=12.20.0
azure-data-tables>=12.5.0
python-dotenv>=1.0.0
tzdata
pytz